refactor verifier and updated utils
