+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.0.5 - 2023-11-17
+
+### Changed
+
+- Moved `getIssuerPublicKeyFromWellKnownURI` from `verifier` to `utils`.
+- Removed console.log from the `verifyVCSDJWT` method in the `verifier` class.
+
 ## 0.0.4 - 2023-10-18
 
 ### Changed
