+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@meeco/sd-jwt-vc",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "SD-JWT VC implementation in typescript",
   "scripts": {
     "build": "tsc",
